+++ README.md
@@ -23,6 +23,7 @@
 Make sure you have the following requirements and packages installed:
 
 - Flutter (Channel Stable, version 3.10.1 )
+- [Intl](https://pub.dev/packages/intl)
 - [Providers](https://pub.dev/packages/provider)
 - [Local_auth](https://pub.dev/packages/local_auth) _(not installed yet)_
 - [Shimmer](https://pub.dev/packages/shimmer) _(not installed yet)_
