+++ lib/main.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:presencee/view/splashscreen/splashView.dart';
+import 'package:presencee/theme/contant.dart';
 import 'package:provider/provider.dart';
+import 'package:presencee/view/home/homePage.dart';
+import 'package:presencee/view/auth/login_view.dart';
+import 'package:presencee/view/splashscreen/splashView.dart';
 
 void main() {
   runApp(const MyApp());
@@ -13,11 +16,13 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        fontFamily: "Poppins",
-        primarySwatch: Colors.blue,
+      theme: Theme.of(context).copyWith(
+        colorScheme: const ColorScheme.light(
+          primary: primaryTheme,
+          secondary: componentsColor,
+        ),
       ),
-      home: const SplashScreen(),
+      home: IntroductionScreen(),
     );
   }
 }
