+++ lib/theme/contant.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+const Color primaryTheme = Color(0xFFFE7968);
+const Color componentsColor = Color.fromRGBO(217, 217, 217, 1);
+const Color highlightSearch = Color.fromRGBO(227, 109, 109, 1);
+const Color iconGray = Color(0xFFD1D1D1);
